Split weather into manageable files
