+++ src/dashboard/weather.py
@@ -5,74 +5,25 @@ import time
 from datetime import datetime
 from typing import List
 
-import altair as alt
 import pandas as pd
 import streamlit as st
 
 # User imports.
-from dashboard import logger
-from dashboard.utils import get_weekends
 from log_keeper.get_config import Database
 from log_keeper.output import weather_to_db
 from log_keeper.weather import WeatherFetcher
-
-weather_metadata = {
-    "wind_speed_10m": {
-        "description": "Wind speed at 10 meters above ground level.",
-        "unit": "kts",
-        "display_name": "Wind Speed",
-    },
-    "wind_direction_10m": {
-        "description": "Wind direction at 10 meters above ground level.",
-        "unit": "degrees",
-        "display_name": "Wind Direction",
-    },
-    "wind_gusts_10m": {
-        "description": "Wind gusts at 10 meters above ground level.",
-        "unit": "kts",
-        "display_name": "Wind Gusts",
-    },
-    "cloud_base_ft": {
-        "description": "Cloud base height in feet.",
-        "unit": "ft",
-        "display_name": "Cloud Base",
-    },
-    "cloud_cover_low": {
-        "description": "Low cloud cover percentage.",
-        "unit": "%",
-        "display_name": "Low Cloud Cover",
-    },
-    "precipitation": {
-        "description": "Precipitation amount.",
-        "unit": "mm",
-        "display_name": "Precipitation",
-    },
-    "temperature_2m": {
-        "description": "Temperature at 2 meters above ground level.",
-        "unit": "°C",
-        "display_name": "Temperature",
-    },
-    "dew_point_2m": {
-        "description": "Dew point temperature at 2 meters above ground level.",
-        "unit": "°C",
-        "display_name": "Dew Point",
-    },
-    "relative_humidity_2m": {
-        "description": "Relative humidity at 2 meters above ground level.",
-        "unit": "%",
-        "display_name": "Relative Humidity",
-    },
-    "surface_pressure": {
-        "description": "Surface pressure.",
-        "unit": "hPa",
-        "display_name": "Surface Pressure",
-    },
-    "datetime": {
-        "description": "Date and time of the weather observation (Local).",
-        "unit": "",
-        "display_name": "Date",
-    },
-}
+from dashboard import logger
+from dashboard.utils import get_weekends
+from dashboard.weather_plots import (
+    plot_launch_vs_nonlaunch_weather,
+    plot_weather_vs_flight_time,
+    plot_wind_polar,
+    weather_table
+)
+from dashboard.weather_utils import (
+    calculate_cloud_base,
+    weather_metadata,
+)
 
 
 def weather_page(db: Database, launches_df: pd.DataFrame):
@@ -145,10 +96,7 @@ def weather_page(db: Database, launches_df: pd.DataFrame):
     # Display the weather data.
     if weather_df is not None:
         # Plot weather vs launches
-        st.subheader("Wind Impact on Flying")
-        plot_wind_vs_launches(weather_df, launches_df)
-
-        st.subheader("Weather Impact on Flight Time")
+        st.subheader("Weather Impact on Flying")
         selected_metric = select_metric_to_plot()
         plot_weather_vs_flight_time(weather_df, launches_df, selected_metric)
         st.subheader("Weather on Launch vs Non-Launch Days")
@@ -157,69 +105,11 @@ def weather_page(db: Database, launches_df: pd.DataFrame):
         )
 
         # Display the weather data in a table.
-        st.subheader("Weather Data")
+        weather_table(weather_df)
 
-        # Use the timezone from the data itself for display formatting
-        display_df = weather_df.copy()
-        display_df["datetime"] = display_df["datetime"].dt.tz_localize(None)
-        # Sort by descending datetime
-        display_df = display_df.sort_values(
-            by="datetime", ascending=False
-        ).reset_index(drop=True)
-
-        st.dataframe(
-            display_df,
-            use_container_width=True,
-            hide_index=True,
-            column_order=[
-                "datetime",
-                "wind_direction_10m",
-                "wind_speed_10m",
-                "wind_gusts_10m",
-                "cloud_base_ft",
-                "cloud_cover_low",
-                "precipitation",
-                "temperature_2m",
-                "dew_point_2m",
-                "relative_humidity_2m",
-                "surface_pressure",
-            ],
-            column_config={
-                "datetime": st.column_config.DatetimeColumn(
-                    "Date", format="DD MMM YY HH:mm (ddd)", pinned=True
-                ),
-                "temperature_2m": st.column_config.NumberColumn(
-                    "Temperature", format="%.1f °C"
-                ),
-                "relative_humidity_2m": st.column_config.NumberColumn(
-                    "Relative Humidity", format="%.1f %%"
-                ),
-                "dew_point_2m": st.column_config.NumberColumn(
-                    "Dew Point", format="%.1f °C"
-                ),
-                "precipitation": st.column_config.NumberColumn(
-                    "Precipitation", format="%.1f mm"
-                ),
-                "surface_pressure": st.column_config.NumberColumn(
-                    "Surface Pressure", format="%.0f hPa"
-                ),
-                "cloud_cover_low": st.column_config.NumberColumn(
-                    "Low Cloud", format="%.0f %%"
-                ),
-                "cloud_base_ft": st.column_config.NumberColumn(
-                    "Cloud Base", format="%.0f ft"
-                ),
-                "wind_speed_10m": st.column_config.NumberColumn(
-                    "Wind Speed", format="%.1f kts"
-                ),
-                "wind_direction_10m": st.column_config.NumberColumn(
-                    "Wind Direction", format="%.0f °"
-                ),
-                "wind_gusts_10m": st.column_config.NumberColumn(
-                    "Wind Gusts", format="%.1f kts"
-                ),
-            },
-        )
+        # Plot wind direction as a polar plot.
+        st.subheader("Wind Direction")
+        plot_wind_polar(weather_df)
 
 
 @st.cache_data(ttl="1h", show_spinner=False)
@@ -347,33 +237,6 @@ def get_weather_data(
     return weather_df
 
 
-def calculate_cloud_base(weather_df: pd.DataFrame) -> pd.DataFrame:
-    """Calculate cloud base from weather data.
-
-    Args:
-        weather_df (pd.DataFrame): DataFrame containing the weather data.
-
-    Returns:
-        pd.DataFrame: DataFrame with cloud base calculated."""
-    # Check if the required columns are present
-    cols_present = all(
-        col in weather_df.columns for col in ["temperature_2m", "dew_point_2m"]
-    )
-    if not cols_present:
-        st.warning("Missing required columns for cloud base calculation.")
-        return weather_df
-
-    # Calculate cloud base using spread between temperature and dew point
-    # The formula is: spread * 400 feet per degree C
-    # (temperature - dew_point) * 400 gives cloud base in feet
-    spread = weather_df["temperature_2m"] - weather_df["dew_point_2m"]
-
-    # Round down to nearest 100 feet.
-    cloud_base = (spread * 400).apply(lambda x: int(x // 100) * 100)
-    weather_df["cloud_base_ft"] = cloud_base
-    return weather_df
-
-
 def get_api_weather_data(
     db: Database, dates: List[datetime.date], show_progress: bool = True
 ) -> pd.DataFrame:
@@ -452,363 +315,6 @@ def get_api_weather_data(
             st.warning("No weather data available for the selected dates.")
         return pd.DataFrame()
     return weather_df
-
-
-def plot_wind_vs_launches(weather_df, launches_df):
-    """Create a scatter plot showing correlation between wind speed
-    and number of launches.
-
-    Args:
-        weather_df (pd.DataFrame): Weather data
-        launches_df (pd.DataFrame): Launches data
-    """
-    # Filter by operating hours for each day.
-    START_HOUR = 9
-    END_HOUR = 17
-    weather_df = weather_df[
-        (weather_df["datetime"].dt.hour >= START_HOUR)
-        & (weather_df["datetime"].dt.hour < END_HOUR)
-    ]
-
-    # Aggregate launches by day
-    daily_launches = (
-        launches_df.groupby(launches_df["Date"].dt.date)
-        .size()
-        .reset_index(name="Launches")
-    )
-
-    # Get average daily wind speed
-    daily_wind = (
-        weather_df.groupby(weather_df["datetime"].dt.date)
-        .agg({"wind_speed_10m": "mean", "wind_gusts_10m": "mean"})
-        .reset_index()
-    )
-
-    # Merge the data
-    merged_df = daily_launches.merge(
-        daily_wind, left_on="Date", right_on="datetime", how="inner"
-    )
-
-    # Create a scatter plot with Altair
-    scatter = (
-        alt.Chart(merged_df)
-        .mark_circle(size=60)
-        .encode(
-            x=alt.X("wind_speed_10m:Q", title="Average Wind Speed (kts)"),
-            y=alt.Y("Launches:Q", title="Number of Launches"),
-            color=alt.Color("wind_gusts_10m:Q", title="Wind Gusts (kts)"),
-            tooltip=[
-                "datetime",
-                "Launches",
-                "wind_speed_10m",
-                "wind_gusts_10m",
-            ],
-        )
-        .properties(
-            title="Wind Speed vs Number of Launches", width=600, height=400
-        )
-    )
-
-    # Add a trend line
-    trend = scatter.transform_regression(
-        "wind_speed_10m", "Launches"
-    ).mark_line(color="red")
-
-    # Display the chart
-    st.altair_chart(scatter + trend, use_container_width=True)
-
-
-def plot_weather_vs_flight_time(weather_df, launches_df, selected_metric):
-    """Create a scatter plot showing how weather affects flight duration.
-
-    Args:
-        weather_df (pd.DataFrame): Weather data
-        launches_df (pd.DataFrame): Launches data
-        selected_metric (str): The weather metric to plot
-    """
-    # Filter by operating hours for each day.
-    START_HOUR = 9
-    END_HOUR = 17
-    weather_df = weather_df[
-        (weather_df["datetime"].dt.hour >= START_HOUR)
-        & (weather_df["datetime"].dt.hour < END_HOUR)
-    ]
-
-    # Get median flight and number of launches per day
-    daily_launches = (
-        launches_df.groupby(launches_df["Date"].dt.date)
-        .agg(
-            {
-                "TakeOffTime": "count",
-                "FlightTime": "median",
-            }
-        )
-        .reset_index()
-        .rename(
-            columns={
-                "TakeOffTime": "Launches",
-                "FlightTime": "MedianFlightTime",
-            },
-        )
-    )
-
-    # Get median daily weather conditions
-    daily_weather = (
-        weather_df.groupby(weather_df["datetime"].dt.date)
-        .agg(
-            {
-                "wind_speed_10m": "median",
-                "wind_gusts_10m": "median",
-                "wind_direction_10m": "median",
-                "cloud_base_ft": "median",
-                "cloud_cover_low": "median",
-                "precipitation": "sum",
-                "temperature_2m": "median",
-                "relative_humidity_2m": "median",
-                "surface_pressure": "median",
-                "dew_point_2m": "median",
-            }
-        )
-        .reset_index()
-    )
-
-    # Merge the data
-    merged_df = daily_launches.merge(
-        daily_weather, left_on="Date", right_on="datetime", how="inner"
-    )
-
-    # Lookup the selected metric in the metadata.
-    metric_display_name = weather_metadata[selected_metric]["display_name"]
-
-    # Append unit to the display name.
-    unit = weather_metadata[selected_metric]["unit"]
-    metric_display_name += f" ({unit})"
-
-    # Create a scatter plot with Altair
-    tooltips = [col for col in merged_df.columns if col != "datetime"]
-
-    # Set specific domain for surface pressure if that's the selected metric
-    x_encoding = alt.X(f"{selected_metric}:Q", title=metric_display_name)
-
-    # Set domain range specifically for surface pressure
-    if selected_metric == "surface_pressure":
-        x_encoding = alt.X(
-            f"{selected_metric}:Q",
-            title=metric_display_name,
-            scale=alt.Scale(domain=[950, 1050]),
-        )
-
-    scatter = (
-        alt.Chart(merged_df)
-        .mark_circle()
-        .encode(
-            x=x_encoding,
-            y=alt.Y("Launches:Q", title="Number of Launches"),
-            size=alt.Size(
-                "MedianFlightTime:Q",
-                scale=alt.Scale(range=[20, 100]),
-                legend=alt.Legend(title="Flight Time (mins)"),
-            ),
-            color=alt.Color("MedianFlightTime:Q"),
-            tooltip=tooltips,
-        )
-        .properties(
-            title=f"{metric_display_name} vs Number of Launches",
-            width=600,
-            height=400,
-        )
-    )
-
-    # Add a trend line for number of launches
-    trend = scatter.transform_regression(
-        f"{selected_metric}", "Launches"
-    ).mark_line(color="red")
-
-    # Display the chart
-    st.altair_chart(scatter + trend, use_container_width=True)
-
-
-def plot_launch_vs_nonlaunch_weather(weather_df, launches_df, selected_metric):
-    """Compare weather conditions on days with launches vs
-    days without launches.
-
-    Args:
-        weather_df (pd.DataFrame): Weather data
-        launches_df (pd.DataFrame): Launches data
-        selected_metric (str): Selected weather metric to compare.
-    """
-    # Number of launches per day cutoff.
-    MIN_LAUNCHES = 10
-
-    # Filter by operating hours for each day.
-    START_HOUR = 9
-    END_HOUR = 17
-    weather_df = weather_df[
-        (weather_df["datetime"].dt.hour >= START_HOUR)
-        & (weather_df["datetime"].dt.hour < END_HOUR)
-    ]
-
-    # Get all weekends.
-    all_weekends = get_weekends(
-        start_date=launches_df["Date"].min(),
-        end_date=launches_df["Date"].max(),
-    )
-
-    # Group launches by date.
-    daily_launches = (
-        launches_df.groupby(launches_df["Date"].dt.date)
-        .agg(
-            {
-                "TakeOffTime": "count",
-                "FlightTime": "median",
-            }
-        )
-        .reset_index()
-        .rename(
-            columns={
-                "TakeOffTime": "Launches",
-                "FlightTime": "MedianFlightTime",
-            },
-        )
-    )
-
-    # Place 0 for weekends with no launches while retaining midweek launches
-    weekend_dates_set = set(all_weekends)
-    launch_dates_set = set(daily_launches["Date"])
-    missing_weekends = list(weekend_dates_set - launch_dates_set)
-
-    if missing_weekends:
-        missing_df = pd.DataFrame(
-            {
-                "Date": missing_weekends,
-                "Launches": [0] * len(missing_weekends),
-                "MedianFlightTime": [0] * len(missing_weekends),
-            }
-        )
-        daily_launches = pd.concat(
-            [daily_launches, missing_df], ignore_index=True
-        )
-
-    # Sort the data by date.
-    daily_launches = daily_launches.sort_values(by="Date").reset_index(
-        drop=True
-    )
-
-    # Get median daily weather conditions
-    daily_weather = (
-        weather_df.groupby(weather_df["datetime"].dt.date)
-        .agg(
-            {
-                "wind_speed_10m": "median",
-                "wind_gusts_10m": "median",
-                "wind_direction_10m": "median",
-                "cloud_base_ft": "median",
-                "cloud_cover_low": "median",
-                "precipitation": "sum",
-                "temperature_2m": "median",
-                "relative_humidity_2m": "median",
-                "surface_pressure": "median",
-                "dew_point_2m": "median",
-            }
-        )
-        .reset_index()
-    )
-
-    # Merge the data
-    merged_df = daily_launches.merge(
-        daily_weather, left_on="Date", right_on="datetime", how="inner"
-    )
-
-    # Split into days with and without launches
-    days_with_launches = merged_df[merged_df["Launches"] >= MIN_LAUNCHES]
-    days_low_launches = merged_df[merged_df["Launches"] < MIN_LAUNCHES]
-
-    # Show how many days with launches vs low launches
-    col1, col2 = st.columns(2)
-    with col1:
-        st.metric(label="Flying days", value=len(days_with_launches))
-    with col2:
-        st.metric(
-            label=f"Poor flying days: (<{MIN_LAUNCHES} launches): ",
-            value=len(days_low_launches),
-        )
-
-    # Get the display name for the selected metric
-    metric_display_name = weather_metadata[selected_metric]["display_name"]
-    metric_unit = weather_metadata[selected_metric]["unit"]
-
-    # Create a DataFrame for plotting
-    plot_data = pd.DataFrame(
-        {
-            "Weather Condition": [metric_display_name] * len(merged_df),
-            "Value": merged_df[selected_metric],
-            "Flying Day": merged_df["Launches"].apply(
-                lambda x: "Launch Day" if x >= MIN_LAUNCHES else "Low Launches"
-            ),
-        }
-    )
-
-    # Set specific domain for surface pressure if that's the selected metric
-    x_encoding = alt.X(
-        "Value:Q",
-        title=metric_display_name,
-    )
-
-    # Set domain range specifically for surface pressure
-    if selected_metric == "surface_pressure":
-        x_encoding = alt.X(
-            "Value:Q",
-            title=f"{metric_display_name} ({metric_unit})",
-            scale=alt.Scale(domain=[950, 1050]),
-        )
-
-    # Create a box plot comparing the metric between days with and
-    # low launches
-    chart = (
-        alt.Chart(plot_data)
-        .mark_boxplot(
-            extent="min-max",
-            ticks=True,
-            median={"color": "black"},
-            rule={"color": "gray"},
-            box={"color": "gray"},
-            outliers={"color": "red"},
-        )
-        .encode(
-            y=alt.Y("Flying Day:N", title=None),
-            x=x_encoding,
-            color=alt.Color("Flying Day:N", legend=None),
-        )
-        .properties(
-            title=f"Comparison of {metric_display_name} on Launch vs "
-            "Low-Launch Days",
-            width=600,
-            height=400,
-        )
-    )
-
-    # Display the chart
-    st.altair_chart(chart, use_container_width=True)
-
-    # Show statistical summary
-    st.subheader("Statistical Summary")
-
-    if len(days_with_launches) > 0 and len(days_low_launches) > 0:
-        launch_median = days_with_launches[selected_metric].median()
-        low_launch_median = days_low_launches[selected_metric].median()
-
-        col1, col2 = st.columns(2)
-        with col1:
-            st.metric(
-                f"Median {metric_display_name} on Launch Days",
-                f"{launch_median:.0f} {metric_unit}",
-            )
-        with col2:
-            st.metric(
-                f"Median {metric_display_name} on Low-Launch Days",
-                f"{low_launch_median:.0f} {metric_unit}",
-                f"{low_launch_median - launch_median:.0f} {metric_unit}",
-            )
 
 
 def select_metric_to_plot():
